Fix nemesis-igmp to allow 0.0.0.0 group address
With IGMP, it should be possible to send a General Query.

http://www.faqs.org/rfcs/rfc2236.html:
   In a Membership Query message, the group address field is set to zero
   when sending a General Query, and set to the group address being
   queried when sending a Group-Specific Query.

This patch is one way to fix this, localized to IGMP.  The problem was
that nemesis_name_resolve() returns -1 when given the IP 0.0.0.0, but in
this context, it is valid.  Another is to change nemesis_name_resolve()
itself to be less picky because it might be useful in other protocols to
set an IP field all zeros.

Signed-off-by: Randy Robertson <rmrobert@vmware.com>
Signed-off-by: Joachim Nilsson <troglobit@gmail.com>