ospf: Default to area 0, and allow setting area 0 as well
The nemesis_name_resolve() function does not allow 0.0.0.0, which
would be quite counter intuitive for a user wanting to inject an
OSPF packet for area 0.

This patch also introduces the ip2int() function.  It returns the
uin32_t representation of an IPv4 network address, in host byte
order format -- because libnet does most of the htonl() for us.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>