configure: Clean up and simplify, a lot
This patch cleans up and simplifies the coonfigure script a lot.  The
purpose is to ease maintenance, remove bugs, and line up the Nemesis
build system along with established best practice.

* Remove home-grown libnet search for library and include files.

The recommended way to provide GNU configure scripts with non-standard
paths to headers and libraries are:

    configure LDFLAGS=-L/path/to/lib CPPFLAGS=-I/path/to/header

* Remove m4/ax_compare_version.m4
* Remove home-grown libnet version check

Instead we now do like arping and other projects do, use standardized
configure checks for APIs only in libnet 1.0.x and libnet 1.1.x.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>