Remove IPv4 address cmp with 0.0.0.0, this is a professional tool
Nemesis users are hard core networking freaks, they know what they're
doing and should be hampered by artificial restrictions like this.

For instance, it should be possible to send an IGMP query with source
address 0.0.0.0, i.e. a proxy query.  It should also be possible to send
general queries, i.e. with group 0.0.0.0.  Which was specifically fixed
in an earlier commit, and we now can revert.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>